requirements.txt
fastapi
requests
yara-python
uvicorn
streamlit
requests
python-dotenv